Modify initial values
This commit changes the value initially present in YAML files in order to match
our needs: change the user name, core and board types.

Also, the sketch itself has been changed in order to test whether the action
will be triggered on pushing the new veresion to 'master' branch of the remote
repository.